Add `RefineFromSpaceField` in order to remove `queryFieldDef` from `ColumnSpaceField`
