+++ packages/malloy/src/lang/ast/field-space/column-space-field.ts
@@ -21,12 +21,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import type {
-  TypeDesc,
-  AtomicFieldDef,
-  QueryFieldDef,
-} from '../../../model/malloy_types';
-import type {FieldSpace} from '../types/field-space';
+import type {TypeDesc, AtomicFieldDef} from '../../../model/malloy_types';
 
 import {SpaceField} from '../types/space-field';
 
@@ -36,10 +31,6 @@ export class ColumnSpaceField<T extends AtomicFieldDef> extends SpaceField {
   }
 
   fieldDef(): T {
-    return this.sourceFieldDef;
-  }
-
-  getQueryFieldDef(_fs: FieldSpace): QueryFieldDef | undefined {
     return this.sourceFieldDef;
   }
